pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["mysubsnewslivecom <my.subs.news@live.com>"]
readme = "README.md"
packages = [{include = "etl", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
psycopg2 = "^2.9.9"
redis = "^5.0.1"
requests = "^2.31.0"
pydantic-settings = "^2.3.4"
apache-airflow-providers-hashicorp = "^3.7.1"
psycopg2-binary = "^2.9.9"
apache-airflow = "^2.9.3"
apache-airflow-providers-postgres = "^5.11.2"
hvac = "^2.3.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pylint = "^3.2.5"
black = "^24.4.2"
ruff = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
